% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs_add_unit_column.R
\name{cbs_add_unit_column}
\alias{cbs_add_unit_column}
\title{For each topic column add a unit column}
\usage{
cbs_add_unit_column(x, columns = colnames(x), ...)
}
\arguments{
\item{x}{\code{data.frame} retrieved using \code{\link[=cbs_get_data]{cbs_get_data()}}.}

\item{columns}{\code{character} with the names of the columns for which labels will be added}

\item{...}{not used.}
}
\value{
the original data.frame \code{x} with extra unit
columns. (see description)
}
\description{
Adds extra unit columns to the dataset that was retrieved using \code{\link[=cbs_get_data]{cbs_get_data()}}.
}
\details{
The unit columns will be named \verb{<topic_column>_unit}, and are a \code{character}

By default all topic columns will be with a unit column. The name
of each unit column will be \verb{<code_column>_unit}.
}
\seealso{
Other data retrieval: 
\code{\link{cbs_add_date_column}()},
\code{\link{cbs_add_label_columns}()},
\code{\link{cbs_download_data}()},
\code{\link{cbs_extract_table_id}()},
\code{\link{cbs_get_data}()},
\code{\link{cbs_get_data_from_link}()}

Other meta data: 
\code{\link{cbs_add_date_column}()},
\code{\link{cbs_add_label_columns}()},
\code{\link{cbs_download_meta}()},
\code{\link{cbs_get_meta}()}
}
\concept{data retrieval}
\concept{meta data}
