% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs_as_sf.R
\name{cbs_as_sf}
\alias{cbs_as_sf}
\title{Create a sf dataframe}
\usage{
cbs_as_sf(x, ...)
}
\arguments{
\item{x}{\code{data.frame} downloaded with \code{\link{cbs_get_data}}.}

\item{...}{reserved for future use.}
}
\description{
Create a sf data.frame
}
