% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs_default_selection.R
\name{cbs_default_selection}
\alias{cbs_default_selection}
\title{extract the default selection from a cbsodata meta object}
\usage{
cbs_default_selection(x, ...)
}
\arguments{
\item{x}{meta object}

\item{...}{for future use}
}
\description{
extract the default selection from a cbsodata meta object
}
