% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs_extract_table_id.R
\name{cbs_extract_table_id}
\alias{cbs_extract_table_id}
\title{extract the id of a cbs table from the statline url}
\usage{
cbs_extract_table_id(url, ...)
}
\arguments{
\item{url}{\code{character} with hyperlink to StatLine table}

\item{...}{future use.}
}
\value{
\code{character} with id, will be NA if not found.
}
\description{
extract the id of a cbs table from the statline url
}
\seealso{
Other data retrieval: 
\code{\link{cbs_add_date_column}()},
\code{\link{cbs_add_label_columns}()},
\code{\link{cbs_add_unit_column}()},
\code{\link{cbs_download_data}()},
\code{\link{cbs_get_data}()},
\code{\link{cbs_get_data_from_link}()}
}
\concept{data retrieval}
