% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs_get_data.R
\name{cbs_get_data}
\alias{cbs_get_data}
\title{Get data from Statistics Netherlands (CBS)}
\usage{
cbs_get_data(id, ..., add_codes = "", recode = FALSE, typed = TRUE,
  add_column_labels = TRUE, add_labels = NULL, dir = tempdir(),
  verbose = FALSE, base_url = CBSOPENDATA, include_ID = FALSE)
}
\arguments{
\item{id}{Identifier of table, can be found in \code{\link{get_table_list}}}

\item{...}{optional filter statemenets}

\item{typed}{Should the data automatically be converted into integer and numeric?}

\item{add_column_labels}{Should column titles be added as a label (TRUE) which are visible in \code{View}}

\item{dir}{Directory where the table should be downloaded. Defaults to temporary
directory}

\item{verbose}{Print extra messages what is happening.}

\item{base_url}{optionally specify a different server. Useful for
third party data services implementing the same protocol.}

\item{include_ID}{Should the data include the ID column for the rows?}
}
\value{
\code{data.frame} with the requested data. Note that a csv copy of the data is stored in \code{dir}.
}
\description{
Retrieves data from a table of Statistics Netherlands. A list of tables
can be retrieved with \code{\link{get_cbs_toc}}. 
Optionaly the data can be filtered on category values. 
The filter is specified with \code{<column_name> = <values>} in which \code{<values>} is a character vector.
Rows with values that are not part of the character vector are not returned. Note that the values
have to be raw (un-recoded) values.
}
\note{
All data are downloaded using \code{\link{download_table}}
}
\examples{
\dontrun{

# get data for main (000000) Consumer Price Index (7196ENG) for March 2000, 
 cbs_get_data(id="7196ENG", Periods="2000MM03", CPI="000000")
}
}
