% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs_get_maps.R
\name{cbs_get_maps}
\alias{cbs_get_maps}
\title{Get list of cbs maps}
\usage{
cbs_get_maps(verbose = FALSE, cache = TRUE)
}
\arguments{
\item{verbose}{if \code{TRUE} a message with the download url will be printed.}

\item{cache}{if \code{TRUE} the result will be cached.}
}
\value{
\code{data.frame} with region, year and links to geojson
}
\description{
Returns a list of (simplified) maps, that can be used with CBS data.
}
\examples{
if (interactive()){
  
  # retrieve maps
  cbs_maps <- cbs_get_maps()
  cbs_maps |> head(4)
  
  gemeente_map <- cbs_get_sf("gemeente", 2023, verbose=TRUE)
  
  # sf object
  gemeente_map
  
  # senseless plot
  plot(gemeente_map, max.plot = 1)
}
}
\seealso{
Other map: 
\code{\link{cbs_get_sf}()}
}
\concept{map}
