% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs_get_sf.R
\name{cbs_get_sf}
\alias{cbs_get_sf}
\title{Retrieve a polygon file}
\usage{
cbs_get_sf(region, year, add_regios = TRUE, verbose = FALSE)
}
\arguments{
\item{region}{\code{character} name of region}

\item{year}{\code{integer} year of a region}

\item{add_regios}{if \code{TRUE} the \verb{$statcode} column is copied to a \verb{$RegioS} column}

\item{verbose}{if \code{TRUE} the method is verbose}
}
\value{
\code{\link[sf:sf]{sf::sf()}} object with the polygons of the regions specified.
}
\description{
Retrieve a polygon sf object.
}
\examples{
if (interactive()){
  
  # retrieve maps
  cbs_maps <- cbs_get_maps()
  cbs_maps |> head(4)
  
  gemeente_map <- cbs_get_sf("gemeente", 2023, verbose=TRUE)
  
  # sf object
  gemeente_map
  
  # senseless plot
  plot(gemeente_map, max.plot = 1)
}
}
\seealso{
Other map: 
\code{\link{cbs_get_maps}()}
}
\concept{map}
