% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs_get_themes.R
\name{cbs_get_tables_themes}
\alias{cbs_get_tables_themes}
\title{Get a the list of tables connected to themes}
\usage{
cbs_get_tables_themes(..., select = NULL, verbose = FALSE, cache = TRUE,
  base_url = CBSOPENDATA)
}
\arguments{
\item{...}{Use this to add a filter to the query e.g. \code{get_tables_themes(ID=10)}.}

\item{select}{\code{character} vector with names of wanted properties. default is all}

\item{verbose}{Print extra messages what is happening.}

\item{cache}{Should the result be cached?}

\item{base_url}{optionally specify a different server. Useful for
third party data services implementing the same protocal.}
}
\value{
A \code{data.frame} with various properties of SN/CBS themes.
}
\description{
Get a the list of tables connected to themes
}
