% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs_search.R
\name{cbs_search}
\alias{cbs_search}
\title{Find tables containing certain words}
\usage{
cbs_search(query, type = c("toc", "docs", "raw"), verbose = FALSE, ...)
}
\arguments{
\item{query}{\code{character} with the words to search for.}

\item{type}{format in which the result should be returned, see details}

\item{verbose}{\code{logical} should the communication with the server be shown?}

\item{...}{not used}
}
\description{
Find tables containing certain words.
}
\details{
The \code{type} can be either:
\itemize{
\item \code{toc}: the same format as \code{\link[=cbs_get_toc]{cbs_get_toc()}}
\item \code{docs}: the table results from the solr query,
\item \code{raw}: the complete results from the solr query.
}
}
