% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substringof.R
\name{contains}
\alias{contains}
\alias{substringof}
\title{Detect substring in column \code{key}}
\usage{
contains(x, key = NULL)

substringof(x, key = NULL)
}
\arguments{
\item{x}{substring to be detected in column}

\item{key}{column name}
}
\description{
Detects a substring in a column.
}
