% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-data.R
\name{download_data-deprecated}
\alias{download_data-deprecated}
\alias{download_data}
\title{Gets all data from a cbs table.}
\usage{
download_data(
  id,
  path = file.path(id, "data.csv"),
  ...,
  select = NULL,
  typed = FALSE,
  verbose = TRUE,
  base_url = getOption("cbsodataR.base_url", BASE_URL)
)
}
\arguments{
\item{id}{of cbs open data table}

\item{path}{of data file, defaults to "id/data.csv"}

\item{...}{optional filter statements to select rows of the data,}

\item{select}{optional names of columns to be returned.}

\item{typed}{Should the data automatically be converted into integer and numeric?}

\item{verbose}{show the underlying downloading of the data}

\item{base_url}{optionally specify a different server. Useful for
third party data services implementing the same protocol.}
}
\description{
This method is deprecated in favor of \code{\link[=cbs_download_data]{cbs_download_data()}}.
}
\seealso{
Other download: 
\code{\link{cbs_download_meta}()},
\code{\link{cbs_download_table}()}

Other data retrieval: 
\code{\link{cbs_add_date_column}()},
\code{\link{cbs_add_label_columns}()},
\code{\link{cbs_add_unit_column}()},
\code{\link{cbs_extract_table_id}()},
\code{\link{cbs_get_data}()},
\code{\link{cbs_get_data_from_link}()}
}
