% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-data.R
\name{download_data}
\alias{download_data}
\title{Gets all data from a cbs table.}
\usage{
download_data(id, path = file.path(id, "data.csv"), ..., select = NULL,
  base_url = CBSOPENDATA)
}
\arguments{
\item{id}{of cbs open data table}

\item{path}{of data file, defaults to "<id>/data.csv"}

\item{...}{optional filter statements to select rows of the data,}

\item{select}{optional names of columns to be returned.}

\item{base_url}{optionally specify a different server. Useful for
third party data services implementing the same protocal.}
}
\description{
Gets all data via bulk download. \code{download_data} dumps the data in (international) csv format.
}
