% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download-meta.R
\name{download_meta}
\alias{download_meta}
\title{Dumps the meta data into a directory}
\usage{
download_meta(id, dir = id, ...)
}
\arguments{
\item{id}{Id of CBS open data table}

\item{dir}{Directory in which data should be stored.
By default it creates a sub directory with the name of the id}

\item{...}{not used}
}
\description{
Dumps the meta data into a directory
}

