% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dump-meta.R
\name{dump_meta}
\alias{dump_meta}
\title{Dumps the meta data into a directory}
\usage{
dump_meta(id, dir = id)
}
\arguments{
\item{id}{Id of CBS open data table}

\item{dir}{Directory in which data should be stored.
By default it creates a sub directory with the name of the id}
}
\description{
Dumps the meta data into a directory
}

