% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{eq}
\alias{eq}
\title{Find codes in columns}
\usage{
eq(x, column = NULL)
}
\arguments{
\item{x}{exact codes to be matched in \code{column}}

\item{column}{name of column}
}
\value{
query object
}
\description{
Find codes in columns
}
\seealso{
Other query: 
\code{\link{has_substring}()}
}
\concept{query}
