% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-meta.R
\name{get_meta_from_dir}
\alias{get_meta_from_dir}
\title{Load meta data from a downloaded table}
\usage{
get_meta_from_dir(dir)
}
\arguments{
\item{dir}{Directory where data was downloaded}
}
\value{
cbs_table object with meta data
}
\description{
Load meta data from a downloaded table
}

