% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-tables.R
\name{get_tables}
\alias{get_tables}
\title{Retrieve a data.frame with requested cbs tables}
\usage{
get_tables(..., select = NULL)
}
\arguments{
\item{...}{filter statement to select rows, e.g. Language="nl"}

\item{select}{\code{character} columns to be returned, by default all columns
will be returned.}
}
\description{
\code{get_tables} by default a list of all tables and all columns will be retrieved.
You can restrict the query by supplying multiple filter statements or by specifying the
columns that should be returned.
}
\details{
\note{get_tables will cache results, so subsequent calls will be much faster.}
}

