% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{has_substring}
\alias{has_substring}
\title{Detect substring in column \code{column}}
\usage{
has_substring(x, column = NULL)
}
\arguments{
\item{x}{substring to be detected in column}

\item{column}{column name}
}
\description{
Detects a substring in a column.
}
\seealso{
Other query: 
\code{\link{eq}()}
}
\concept{query}
