% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substringof.R
\name{key_contains}
\alias{key_contains}
\alias{substringof}
\title{Detect substring in column \code{key}}
\usage{
key_contains(x, key)

substringof(x, key)
}
\arguments{
\item{x}{substring to be detected in column}

\item{key}{column name}
}
\description{
Detects a substring in a column.
}
