% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeplink.R
\name{resolve_deeplink}
\alias{resolve_deeplink}
\title{resolve a deeplink created in the opendata portal}
\usage{
resolve_deeplink(
  deeplink,
  ...,
  base_url = getOption("cbsodataR.base_url", BASE_URL)
)
}
\arguments{
\item{deeplink}{url to the deeplink in the opendataportal}

\item{...}{used in the query}

\item{base_url}{optionally specify a different server. Useful for
third party data services implementing the same protocol.}
}
\value{
information object with table id, select, filter and query statement.
}
\description{
resolve a deeplink created in the opendata portal
}
