% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs_get_data.R
\name{cbs_get_data}
\alias{cbs_get_data}
\title{Get data from Statistics Netherlands (CBS)}
\usage{
cbs_get_data(id, ..., select = NULL, typed = TRUE,
  add_column_labels = TRUE, dir = tempdir(), verbose = FALSE,
  base_url = getOption("cbsodataR.base_url", BASE_URL),
  include_ID = FALSE)
}
\arguments{
\item{id}{Identifier of table, can be found in \code{\link{cbs_get_toc}}}

\item{...}{optional filter statements}

\item{select}{\code{character} optional, columns to select}

\item{typed}{Should the data automatically be converted into integer and numeric?}

\item{add_column_labels}{Should column titles be added as a label (TRUE) which are visible in \code{View}}

\item{dir}{Directory where the table should be downloaded. Defaults to temporary
directory}

\item{verbose}{Print extra messages what is happening.}

\item{base_url}{optionally specify a different server. Useful for
third party data services implementing the same protocol.}

\item{include_ID}{Should the data include the ID column for the rows?}
}
\value{
\code{data.frame} with the requested data. Note that a csv copy of the data is stored in \code{dir}.
}
\description{
Retrieves data from a table of Statistics Netherlands. A list of available tables
can be retrieved with \code{\link{cbs_get_toc}}. Use the \code{Identifier} column of 
\code{cbs_get_toc} as \code{id} in \code{cbs_get_data} and \code{cbs_get_meta}.
Optionaly the data can be filtered on category values. 
The filter is specified with \code{<column_name> = <values>} in which \code{<values>} is a character vector.
Rows with values that are not part of the character vector are not returned. Note that the values
have to be raw (un-recoded) values.
}
\details{
By default the columns will be converted to their type (\code{typed=TRUE}).
CBS uses multiple types of missing (unknown, surpressed, not measured, missing): users
wanting all these nuances can use \code{typed=FALSE} which results in character columns.
}
\note{
All data are downloaded using \code{\link{cbs_download_table}}
}
\examples{
\dontrun{

# get data for main (000000) Consumer Price Index (7196ENG) for March 2000, 
 cbs_get_data(id="7196ENG", Periods="2000MM03", CPI="000000")
}
}
\seealso{
\code{\link{cbs_get_meta}}, \code{\link{cbs_download_data}}
}
