% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs_get_meta.R
\name{cbs_get_meta}
\alias{cbs_get_meta}
\title{Get metadata of a cbs table}
\usage{
cbs_get_meta(id, verbose = FALSE, cache = TRUE,
  base_url = getOption("cbsodataR.base_url", BASE_URL))
}
\arguments{
\item{id}{internal id of CBS table, can be retrieved with \code{\link{cbs_get_toc}}}

\item{verbose}{Print extra messages what is happening.}

\item{cache}{should the result be cached?}

\item{base_url}{optionally specify a different server. Useful for
third party data services implementing the same protocol.}
}
\value{
cbs_table object containing several \code{data.frames} with meta data.
}
\description{
Retrieve the meta data of a CBS open data table.  Caching (\code{cache=TRUE}) improves
the performance considerably. The meta data of a CBS table is
}
