% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{contains}
\alias{contains}
\alias{substringof}
\title{Detect substring in column \code{column}}
\usage{
contains(x, column = NULL)

substringof(x, column = NULL)
}
\arguments{
\item{x}{substring to be detected in column}

\item{column}{column name}
}
\description{
Detects a substring in a column.
}
\seealso{
Other query: 
\code{\link{eq}()}
}
\concept{query}
