% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-meta.R
\name{download_meta}
\alias{download_meta}
\title{Dumps the meta data into a directory}
\usage{
download_meta(id, dir = id, ..., cache = FALSE, base_url = CBSOPENDATA)
}
\arguments{
\item{id}{Id of CBS open data table}

\item{dir}{Directory in which data should be stored. 
By default it creates a sub directory with the name of the id}

\item{...}{not used}

\item{cache}{Should meta data be cached?}

\item{base_url}{optionally allow to specify a different server. Useful for
third party data services implementing the same protocal.}
}
\value{
meta data object
}
\description{
Dumps the meta data into a directory
}

