% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download-table.R
\name{download_table}
\alias{download_table}
\title{Download a table from statistics Netherlands}
\usage{
download_table(id, dir = id, ...)
}
\arguments{
\item{id}{Identifier of CBS table (can be retrieved from \code{\link{get_table_list}})}

\item{dir}{Directory where table should be downloaded}

\item{...}{Parameters passed on to \code{\link{download_data}}}
}
\description{
Download a table from statistics Netherlands
}

