% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dump-data.R
\name{dump_data}
\alias{dump_data}
\title{Gets all data from a cbs table.}
\usage{
dump_data(id, path = file.path(id, "data.csv"), ..., select = NULL)
}
\arguments{
\item{id}{of cbs open data table}

\item{path}{of data file, defaults to "<id>/data.csv"}

\item{...}{optional filter statements to select rows of the data, e.g.}

\item{select}{optional names of columns to be returned.}
}
\description{
Gets all data via bulk download. Dumps the data in (international) csv format.
}

