% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dump-data.R
\name{dump_data}
\alias{dump_data}
\title{Gets all data via bulk download. Dumps the data in (international) csv format.}
\usage{
dump_data(id, path = file.path(id, "data.csv"))
}
\arguments{
\item{path}{of data file, defaults to "<id>/data.csv"}

\item{table}{id of cbs open data table}
}
\description{
Gets all data via bulk download. Dumps the data in (international) csv format.
}

