% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_data}
\alias{get_data}
\title{Get data from Statistics Netherlands (CBS)}
\usage{
get_data(id, ..., recode = TRUE, dir = tempdir())
}
\arguments{
\item{id}{Identifier of table, can be found in \code{\link{get_table_list}}}

\item{...}{optional filter statemenets}

\item{recode}{Should the categories of the table be recoded with their title
or with their key?}

\item{dir}{Directory where the table should be downloaded. Defaults to temporary
directory}
}
\description{
Retrieves data from a table of Statistics Netherlands. A list of tables
can be retrieved with \code{\link{get_table_list}}.
}
\details{
Optionaly the data can be filtered on category values.
e.g.
\code{get_data(id, Period=c('2014', 2015'), Country=c('NL'))}
where \code{Period} and \code{Country} are dimensions of
}
\note{
All data for this table are downloaded using
\code{\link{download_table}}.
}

