% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_data}
\alias{get_data}
\title{Get data from Statistics Netherlands (CBS)}
\usage{
get_data(id, ..., recode = TRUE, dir = tempdir(), base_url = CBSOPENDATA)
}
\arguments{
\item{id}{Identifier of table, can be found in \code{\link{get_table_list}}}

\item{...}{optional filter statemenets}

\item{recode}{Should the categories of the table be recoded with their title
(TRUE) or with their key (FALSE)? Should column names be coded with title (TRUE)
or key (FALSE)}

\item{dir}{Directory where the table should be downloaded. Defaults to temporary
directory}

\item{base_url}{optionally specify a different server. Useful for
third party data services implementing the same protocol.}
}
\description{
Retrieves data from a table of Statistics Netherlands. A list of tables
can be retrieved with \code{\link{get_table_list}}. 
Optionaly the data can be filtered on category values.
}
\note{
All data are downloaded using \code{\link{download_table}}
}

