% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-meta.R
\name{get_meta}
\alias{get_meta}
\title{gets metadata of a cbs table}
\usage{
get_meta(id, cache = FALSE)
}
\arguments{
\item{id}{internal id of CBS table}

\item{cache}{should the result be cached?}
}
\description{
gets metadata of a cbs table
}

