% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-meta.R
\name{get_meta}
\alias{get_meta}
\title{Get metadata of a cbs table}
\usage{
get_meta(id, verbose = TRUE, cache = FALSE, base_url = CBSOPENDATA)
}
\arguments{
\item{id}{internal id of CBS table, can be retrieved with \code{\link{get_table_list}}}

\item{verbose}{Print extra messages what is happening.}

\item{cache}{should the result be cached?}

\item{base_url}{optionally specify a different server. Useful for
third party data services implementing the same protocal.}
}
\value{
cbs_table object containing serveral \code{data.frames} with meta data.
}
\description{
Retrieve the meta data of a CBS open data table.  Caching (\code{cache=TRUE}) may improve
the performance considerably. The meta data of a CBS table is
}
