% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-tables.R
\name{get_tables}
\alias{get_tables}
\title{gets list of all cbs tables}
\usage{
get_tables(..., select = NULL)
}
\arguments{
\item{...}{filter statement to select rows, e.g. Language="nl"}

\item{select}{\code{character} columns to be returned, by default all columns
will be returned.}
}
\description{
gets list of all cbs tables
}

