% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-themes.R
\name{get_themes}
\alias{get_themes}
\title{gets list of all cbs themes}
\usage{
get_themes(..., select = NULL, base_url = CBSOPENDATA)
}
\arguments{
\item{...}{Use this to add a filter to the query e.g. \code{get_themes(ID=10)}}

\item{select}{\code{character} vector with names of wanted properties. default is all}

\item{base_url}{optionally specify a different server. Useful for
third party data services implementing the same protocal.}
}
\description{
Returns a list of all cbs themes.
}

